Create step1-billboard.py
Billboard has an end-of-year top 100 for every year, e.g., https://www.billboard.com/charts/year-end/1970/hot-100-songs. There is a system in the URLs, which makes it easy to automatically request the html of page after page. Per year, I scraped the artist and track titles and stored them in a csv file with three variables: year, artist, and track. Billboard.com is sensitive to rapid consecutive requests, so I paused the script for about 2-3 seconds at the end of each iteration using the Python pause module.